+++ app/src/main/java/com/example/tidy/TaskService.kt
@@ -1,12 +1,10 @@
 package com.example.tidy
 
-import android.content.Context
 import com.tidy.sqldelight.Task
 import com.yourapp.db.AppDatabase
 
 class TaskService(
     db: AppDatabase,
-    private val context: Context,
     private val scheduleService: ScheduleService
 ) : DbOperation(db = db) {
     override suspend fun saveTask(task: Task): Long {
@@ -17,17 +15,14 @@ class TaskService(
             return id
         } else {
             super.updateTask(task)
-            Utils.cancelAlarm(context, task.id)
             if (task.repeatAfterDone == 1L && task.done == 0L) return task.id
-            scheduleService.scheduleTask(task)
+            scheduleService.rescheduleTask(task)
             return task.id
         }
     }
 
     override suspend fun deleteTask(id: Long) {
         super.deleteTask(id)
-        Utils.cancelAlarm(
-            context = context, taskId = id
-        )
+        scheduleService.cancelSchedule(taskId = id)
     }
 }
