+++ app/src/main/java/com/example/tidy/Utils.kt
@@ -36,7 +36,6 @@ import androidx.work.OutOfQuotaPolicy
 import androidx.work.WorkManager
 import androidx.work.workDataOf
 import com.Ace777.tidy.R
-import com.example.tidy.constants.Options
 import com.example.tidy.constants.RepeatTypes
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.gson.Gson
@@ -317,45 +316,6 @@ object Utils {
         private var lastId = 0
         fun nextId(): Int = ++lastId
     }
-
-    fun scheduleAlarm(context: Context, scheduleTime: Long, action: String, taskId: Long) {
-        val alarmManager = context.getSystemService(ALARM_SERVICE) as AlarmManager
-        val intent = Intent(context, AlarmReceiver::class.java).apply {
-            putExtra(Options.ACTION, action)
-            putExtra(Options.TASK_ID, taskId)
-        }
-
-        val pendingIntent = PendingIntent.getBroadcast(
-            context,
-            taskId.toInt(),
-            intent,
-            PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
-        )
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S && !alarmManager.canScheduleExactAlarms()) {
-            println("permission not given")
-            return
-        }
-
-        alarmManager.setExactAndAllowWhileIdle(
-            AlarmManager.RTC_WAKEUP,
-            scheduleTime,
-            pendingIntent
-        )
-    }
-
-    fun cancelAlarm(context: Context, taskId: Long) {
-        val alarmManager = context.getSystemService(ALARM_SERVICE) as AlarmManager
-        val intent = Intent(context, AlarmReceiver::class.java)
-        val pendingIntent = PendingIntent.getBroadcast(
-            context,
-            taskId.toInt(),
-            intent,
-            PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
-        )
-        alarmManager.cancel(pendingIntent)
-    }
-
     fun requestExactAlarmPermission(context: Context) {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
             val alarmManager = context.getSystemService(ALARM_SERVICE) as AlarmManager
